Ayurveda_Chatbot/requirements.txt
Flask==3.0.2
flask-pymongo==2.3.0
google-generativeai==0.7.2
python-dotenv==1.0.1
